requirements.txt
grpcio==1.43.0
grpcio-tools==1.43.0
numpy==1.19.2
protobuf==3.15.8
tensorflow==2.8.0
watchdog==0.10.3
